Rename Rest Service
