+++ project1-react-mid/src/restService/usersService.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import * as serviceData from "./serviceData.js";
+import * as serviceData from "./restService";
 const path = "users";
 
 async function getAllUsers() {
